requirements.txt
streamlit>=1.38.0
easyocr>=1.7.1
Pillow>=10.3.0
numpy>=1.26.4

